+++ MDPlayer/MDPlayer/Driver/MNDRV/devpsgemu.cs
@@ -64,7 +64,7 @@ namespace MDPlayer.Driver.MNDRV
                 reg.D1_L = 0;
             }
 
-            reg.D1_W = (UInt16)(-reg.D1_W);
+            reg.D1_W = (UInt16)(-(Int16)reg.D1_W);
 
             reg.D1_W = (UInt16)((Int16)reg.D1_W >> 2);
             reg.D0_W += (UInt32)(Int16)reg.D1_W;
@@ -331,7 +331,7 @@ namespace MDPlayer.Driver.MNDRV
             reg.D0_W >>= (int)reg.D1_W;
             reg.D1_W = mm.ReadUInt16(reg.a5 + w.detune);
             reg.D1_W += mm.ReadUInt16(reg.a5 + w.addkeycode);
-            reg.D1_W = (UInt16)(-reg.D1_W);
+            reg.D1_W = (UInt16)(-(Int16)reg.D1_W);
             reg.D1_W = (UInt16)((Int16)reg.D1_W >> 2);
             reg.D0_W += (UInt32)(Int16)reg.D1_W;
             _emu_set_psg_bend();
